阶段保存

+++ djs-db/src/main/java/com/valueservice/djs/db/dao/mini/MiniSignMapper.java
@@ -16,7 +16,7 @@ public interface MiniSignMapper {
 
     int updateByPrimaryKey(MiniSign record);
 
-    MiniSign selectByMiniUserId(@Param("miniUserId")Long miniUserId);
+    MiniSign selectByMiniUserId(@Param("miniUserId")Integer miniUserId);
 
 
 }
